improve debug symbol discovery
If elfutils fails to find the debug symbols, the perfparser tries to
find it by checking first in /usr/lib/ then in .debug and then in
/usr/lib/debug. This causes some problems if the debug file has the same
name as the library itself. In this case the algorithm stops at
/usr/lib/ and returns the wrong file.
To prevent this I reorder the algorithm to first check in .debug then in
/usr/lib/debug and then in /usr/lib/

I also changed how the algorithm searches inside /usr/lib/debug. Some
distros (e.g. Ubuntu) uses shorted path names so the debug file for
/usr/lib/x86_64-linux-gnu/libc-2.31.so is located in
/usr/lib/debug/lib/x86_64-linux-gnu/libc-2.31.so. To find the file the
/usr/lib/debug/usr/lib and /usr/lib/debug/lib are searched.

fixes KDAB/hotspot#265
fixes KDAB/hotspot#128

Change-Id: Iab2283819f947e89e9a933e66dfb7f45ff9156e1